feat: Add links and GitHub icons to case studies
This change introduces optional `github` and `link` properties to the `CaseStudy` type and populates them in the `data.ts` file. It also updates the `CaseStudies` component to display GitHub and external link icons and corresponding anchor tags when these properties are present, enhancing user interaction and providing direct access to project details.